auction/requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.8.5
django-filter==2.4.0
markdown==3.3.4
djangorestframework==3.12.4
pillow==8.3.1